Adds test for office and living space allocation
