+++ app/amity.py
@@ -1,32 +1,11 @@
 class Amity(object):
 
-    
-    room_name = ''
-    room_type = ''
-
-
     def load_people(self):
         '''
         Loads people to the system from a file.
         '''
         pass
 
-
-
-    def allocate_office(self):
-        '''
-        Allocates office to the people once they are added to the list
-        at random. Allocates offices that are not full yet.
-        '''
-        pass
-
-    def allocate_living_space(self):
-        '''
-        Allocates the fellows a living space if they want one.
-        Have a default value of No.
-        Allocates guys to the male wing and ladies to the female wing
-        '''
-        pass
 
     def reallocate_person(self, person_id):
         '''
